requirements.txt
atari-py==0.1.1
azure==1.0.3
azure-common==1.1.8
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==2.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
baselines==0.1.4
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
dill==0.2.7.1
enum34==1.1.6
feather-format==0.4.0
future==0.16.0
gym==0.9.4
html5lib==0.9999999
idna==2.6
joblib==0.11
Keras==2.1.2
Markdown==2.6.11
matplotlib==2.1.1
networkx==2.0
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.0.0
progressbar2==3.34.3
protobuf==3.5.1
pyarrow==0.8.0
pyglet==1.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
tqdm==4.19.5
urllib3==1.22
Werkzeug==0.14.1
zmq==0.0.0
